requirements.txt
matplotlib==3.7.0
pandas==1.5.3
Pillow==9.4.0
streamlit==1.19.0
torch==1.13.1
torchvision==0.14.1
